Latest ---

+++ Hld.WebApplication/Hld.WebApplication/Hld.WebApplication/Hld.WebApplication/ViewModel/ZincWatchListSummaryViewModal.cs
@@ -10,8 +10,8 @@ namespace Hld.WebApplication.ViewModel
         public int JobID { get; set; }
         public DateTime? StartTime { get; set; }
         public DateTime? CompletionTime { get; set; }
-        public int Total_ASIN { get; set; }
-        public int Available { get; set; }
+        public decimal Total_ASIN { get; set; }
+        public decimal Available { get; set; }
         public int Prime { get; set; }
         public int NoPrime { get; set; }
         public int Unavailable { get; set; }
